Implemented the user list that fetches data from GraphQL
