+++ frontend/src/App.module.scss
@@ -27,6 +27,8 @@
 
   p {
     font-weight: 300;
+    line-height: $p-line-height;
+    font-size: $p-font-size;
   }
 
 }
