+++ frontend/src/sass/variables.scss
@@ -7,6 +7,10 @@ $sans-pro-font-family: 'Source Sans Pro', $default-font-family;
 
 $page-background-color: #F8F8F8;
 
+//p
+$p-font-size: 16px;
+$p-line-height: 20px;
+
 //H1
 $h1-font-size: 48px;
 $h1-font-weight: 300;
